[refactor] Grouping Strategy Classes into strategy Package.
